+++ src/components/BrandGradientText/index.ts
@@ -1,0 +1,3 @@
+import PBrandGradientText from './BrandGradientText.vue';
+
+export default PBrandGradientText;