+++ src/components/Checkbox/Checkbox.vue
@@ -8,10 +8,10 @@
       class="ph-checkbox ph-absolute ph-opacity-0"
       @change="onChange"
     />
-    <label :for="id" class="ph-text-sm ph-pl-2 ph-transition ph-duration-500 ph-cursor-pointer">
+    <label :for="id" class="ph-pl-2 ph-transition ph-duration-500 ph-cursor-pointer">
       <div
-        class="ph-w-6 ph-h-6 ph-rounded-md ph-border ph-absolute ph-left-0 ph-flex ph-items-center ph-justify-center ph-transition"
-        :class="value ? 'ph-bg-brand2 ph-border-brand2' : 'ph-bg-grey6 ph-border-grey5'"
+        class="ph-checkbox-toggle ph-w-6 ph-h-6 ph-rounded-md ph-border ph-absolute ph-left-0 ph-flex ph-items-center ph-justify-center ph-transition"
+        :class="value ? 'ph-bg-brand2 ph-border-brand2' : 'ph-bg-grey6 '+ darkBorder ? 'ph-border-grey4' : 'ph-border-grey5'"
       >
         <p-icon name="Checkmark" type="xs" class="ph-text-white ph-transition ph-duration-300 ph-transform" :class="value ? 'ph-opacity-1 ph-scale-100' : 'ph-opacity-0 ph-scale-0'"/>
       </div>
@@ -56,6 +56,10 @@ export default Vue.extend({
       type: String as PropType<string>,
       default: '',
     },
+    darkBorder: {
+      type: Boolean as PropType<boolean>,
+      default: false,
+    }
   },
 
   data() {
@@ -85,5 +89,7 @@ export default Vue.extend({
 </script>
 
 <style lang="postcss" scoped>
-
+.ph-checkbox:focus +label .ph-checkbox-toggle{
+  border: 2px solid #3DD6C7!important;
+}
 </style>