+++ src/components/ListItem/ListItem.vue
@@ -1,0 +1,14 @@
+<template>
+  <li :bind="$attrs">
+    <div class="ph-ml-3">
+      <slot />
+    </div>
+  </li>
+</template>
+<script lang='ts'>
+import Vue, { PropType } from 'vue';
+
+export default Vue.extend({
+  name: 'PLi',
+});
+</script>