+++ src/components/ListItem/index.ts
@@ -1,0 +1,3 @@
+import PLi from './ListItem.vue';
+
+export default PLi;