+++ CHANGELOG.md
@@ -6,9 +6,30 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.1.580] - 2021-09-22
+## [0.1.586] - 2021-10-01
+## [0.1.585] - 2021-10-01
 ### Added
-- remove Checkbox focus color, add cursor
+- New SVG for gradient backgrounds
+
+
+
+## [0.1.583] - 2021-09-30
+### Added
+- Resolve Checkbox dark border and font size
+
+## [0.1.582] - 2021-09-30
+### Added
+- Added BrandGradientText and InfoCard
+
+## [0.1.581] - 2021-09-29
+### Added
+- Button component Safari enhancement
+- Change input error display
+- Added List/List item component
+
+## [0.1.580] - 2021-09-27
+### Added
+- Remove Checkbox focus color, add cursor
 
 ## [0.1.579] - 2021-09-24
 ## [0.1.578] - 2021-09-24